requirements.txt
djangorestframework
django-rest-auth